Refactor editorExtension.ts for improved modularity.
- Split editorExtension.ts into smaller, more manageable files:
  - decorationBuilder.ts: Handles building decorations
  - colorDecorationUtils.ts: Manages color decoration specifics
  - syntaxTreeUtils.ts: Contains utility functions for working with syntax trees
  - hoverHandler.ts: Handles hover behavior
- Moved utility functions to their respective files and updated import statements
- Ensured proper passing of ColorPicker instance through relevant functions
